migrate posts to use clubs and refactor poll options
